+++ ORBITER/src/lib/top_level/isomorph/isomorph_global.cpp
@@ -280,7 +280,7 @@ void isomorph_read_solution_files(
 	const char *prefix_iso, int level,
 	const char **fname, int nb_files, 
 	int f_has_final_test_function, 
-	int (*final_test_function)(int *data, int sz,
+	int (*final_test_function)(long int *data, int sz,
 			void *final_test_data, int verbose_level),
 	void *final_test_data, 
 	int verbose_level)
